separate footprint processing function from other scaffolding code which just deals with reading/writing files.
